Payment tracker (#1828)
* changed mdms path

* adding redis configs

* Adding sample response

* Adding sample response as string

* adding report api

---------

+++ backend/wms/src/main/java/org/egov/wms/web/model/Total.java
@@ -15,7 +15,7 @@ import java.util.List;
 public class Total {
 
     @JsonProperty("estimatedAmount")
-    private Integer estimatedAmount;
+    private Double estimatedAmount;
 
     @JsonProperty("paymentDetails")
     private List<PaymentDetailsByBillType> paymentDetailsByBillType;
